整理完毕，有最基本的增删改查操作

+++ mybatis/src/main/java/com/miaxis/entity/Author.java
@@ -4,15 +4,18 @@ import java.util.List;
 
 public class Author {
 
-	private int id;
+	private Long id;
 	private String name;
 	private List<Blog> blogList;
-	public int getId() {
+
+	public Long getId() {
 		return id;
 	}
-	public void setId(int id) {
+
+	public void setId(Long id) {
 		this.id = id;
 	}
+
 	public String getName() {
 		return name;
 	}
